+++ src/main/kotlin/com/github/rebel000/cmdlineargs/ui/ArgumentTreeRootNode.kt
@@ -3,9 +3,13 @@ package com.github.rebel000.cmdlineargs.ui
 import com.github.rebel000.cmdlineargs.ArgumentFilter
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
+import com.intellij.util.ui.ThreeStateCheckBox
 import java.util.*
 
 class ArgumentTreeRootNode(name: String) : ArgumentTreeNode(name, isFolder = true, readonly = true) {
+    override val state: ThreeStateCheckBox.State
+        get() = ThreeStateCheckBox.State.DONT_CARE
+
     override fun getArgs(out: Vector<String>, filter: ArgumentFilter) {
         out.ensureCapacity(out.size + childCount)
         forEachArg {
@@ -38,5 +42,8 @@ class ArgumentTreeRootNode(name: String) : ArgumentTreeNode(name, isFolder = tru
             return true
         }
         return false
+    }
+
+    override fun update() {
     }
 }